backend/requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# OpenAI and LLMs
openai==1.59.3

# RAG Components
chromadb==0.4.22
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Data Processing
pandas==2.1.4
tiktoken==0.5.2

# Document Loading
pypdf==3.17.4
markdown==3.5.1
unstructured==0.11.0

# CORS and middleware
starlette==0.35.1

# Observability (Optional)
langsmith==0.0.77
